pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "catalyst-automation"
version = "0.1.0"
description = "Cisco Catalyst Center automation skeleton (Python)"
requires-python = ">=3.9"
dependencies = [
  "requests>=2.32.0",
  "python-dotenv>=1.0.1",
  "PyYAML>=6.0.2",
  "tenacity>=9.0.0",
  "tabulate>=0.9.0",
]

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100